% Generated by roxygen2 (4.0.2): do not edit by hand
\name{circleFun}
\alias{circleFun}
\title{Extracting paramenters}
\usage{
circleFun(center = c(0, 0), radio = 1, npoints = 100)
}
\arguments{
\item{center}{a list of parameters.}

\item{radio}{the parameter value that will be extracted.}

\item{npoints}{the parameter value that will be extracted.}
}
\value{
The parameter value that was extracted.
}
\description{
Function to extract a parameter value from a list
}
\author{
Valeria Gogni, Mariano Bonoli, Ruben Bufanio, Diego Edwards
}

