% Generated by roxygen2 (4.0.2): do not edit by hand
\name{energy}
\alias{energy}
\title{Conversion of dataframe in class windata}
\usage{
energy(datawd, datawtg, ane, model, dist = NA)
}
\arguments{
\item{datawd}{a dataframe to be converted}

\item{datawtg}{interval of time betwen two registers. Actually, it's only acepted intervals of 1 minut.}

\item{ane}{the name of the variable that contains the dates of measurements}

\item{model}{the admit formats are:}

\item{dist}{the admit formats are:}
}
\value{
Object of class 'windata' (see details).
}
\description{
Shape a dataframe in a class 'windata'
}
\details{
The object windata is a list with the parameters that were mencionated before.
}
\examples{
# simple example using the windspeed data set
data(wd)

# let's examine windspeed to see the variables' names
head(wd)
}
\author{
Valeria Gogni, Mariano Bonoli, Ruben Bufanio, Diego Edwards
}

