% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fitWD}
\alias{fitWD}
\title{Fit differents probability distributions}
\usage{
fitWD(datawd, ane)
}
\arguments{
\item{datawd}{object of class \code{windata}.}

\item{ane}{one character strings with anemometer name to fit.}
}
\value{
A list with the following components:
-aic: Akaike Index Coefficient.

-loglik: maximun likelihood.

-parameters: a list with the following components:
 data: data to estimate.
 name: parameter name.
 type: parameter type.
 value: estimated parameter.
 sd: standard desvest of estimated parameter.
}
\description{
Fit by maximum likelihood Weibull, Gamma and Lognormal distributions.
}
\examples{
data("wd", package = "WindResource")

## Getting anemometer names
wd$ane$ane.names

## Simple example using the wd class
fitWD(wd, ane = "ane10")
}
\author{
Mariano Bonoli Escobar, Diego Edwards, Valeria Gogni, Ruben Bufanio
}

