% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gen10m}
\alias{gen10m}
\title{Conversion of dataframe in class windata}
\usage{
gen10m(wd, Ns = 60)
}
\arguments{
\item{wd}{a dataframe to be converted}

\item{Ns}{interval of time betwen two registers. Actually, it's only acepted intervals of 1 minut.}
}
\description{
Shape a dataframe in a class 'windata'
}
\details{
The object windata is a list with the parameters that were mencionated before.
}
\examples{
# simple example using the windspeed data set
data(wd)

# let's examine windspeed to see the variables' names
head(wd)
}
\author{
Valeria Gogni, Mariano Bonoli, Ruben Bufanio, Diego Edwards
}

