\name{getParam}
\alias{getParam}
\title{Extracting paramenters}
\usage{
  getParam(list, param)
}
\arguments{
  \item{list}{a list of parameters.}

  \item{param}{the parameter value that will be extracted.}
}
\value{
  The parameter value that was extracted.
}
\description{
  Function to extract a parameter value from a list
}
\author{
  Valeria Gogni, Mariano Bonoli, Ruben Bufanio, Diego
  Edwards
}

