% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotcalendar}
\alias{plotcalendar}
\title{Conversion of dataframe in class windata}
\usage{
plotcalendar(wd, var = c("ave", "min", "max"), ane = "Ane1", shiny = F)
}
\arguments{
\item{wd}{a dataframe to be converted}

\item{var}{interval of time betwen two registers. Actually, it's only acepted intervals of 1 minut.}

\item{ane}{the name of the variable that contains the dates of measurements}

\item{shiny}{the admit formats are:
- 'YYYYMMDD','YYYY-MM-DD','YYYY.MM.DD' or 'DD/MM/YYYY'}
}
\value{
Object of class 'windata' (see details).
}
\description{
Shape a dataframe in a class 'windata'
}
\details{
The object windata is a list with the parameters that were mencionated before.
}
\examples{
# simple example using the windspeed data set
data(wd)

# let's examine windspeed to see the variables' names
head(wd)
}
\author{
Valeria Gogni, Mariano Bonoli, Ruben Bufanio, Diego Edwards
}

