% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWD.R
\name{plotWD}
\alias{plotWD}
\title{Plots for \code{Windata} objects}
\usage{
plotWD(datawd, ane = NA, var = NA, type = c("histogram"), by = "none",
  since = NULL, to = NULL, binwidth = 1)
}
\arguments{
\item{datawd}{an object of class \code{windata}.}

\item{ane}{a vector of character strings with anemometers names to plot.}

\item{var}{a vector of character strings the the variables to plot: \code{ave}, \code{min}, \code{max}, \code{freq}.}

\item{type}{type of plot: \code{histogram}, \code{rose}, \code{profiles} and \code{boxplot}. See also Details.}

\item{by}{an optional string stating if the plot is divided in panels by \code{month} or \code{hour}.}

\item{since}{an optional string with initial date to be taken into account to make the plot. The string format is 'YYYY-MM-DD'.}

\item{to}{an optional string indicating final date to be taken into account to make the plot. The string format is 'YYYY-MM-DD'.}

\item{binwidth}{an optional bindwidth interval used in histogram plots.}
}
\description{
The function allows to describe graphically the main features of the current wind speed and direction.
This function generates histograms, correlations, wind roses, profile graphs and boxplots. It also allows to evaluate the fit of the data to probability distributions.
}
\examples{
data("wdMtTom", package = "WindResource")
data("wd10", package = "WindResource")

## Getting anemometer names
wdMtTom$ane$ane.names
wd10$ane$ane.names

## Histogram
plotWD(data=wdMtTom, ane="Anem24bMS", type="histogram", by="none")
plotWD(data=wdMtTom, ane="Anem24bMS", type="histogram", by="none", binwidth=.5)
plotWD(data=wdMtTom, ane="Anem24bMS", type="histogram", by="hour")

## Rose
plotWD (data=wdMtTom, ane=c("Anem24bMS", "Anem37aMS"), var="mean", type="rose", by="none")
plotWD (data=wdMtTom, ane=c("Anem37aMS"), var="mean", type="rose", by="month")

## Boxplot
plotWD (data=wdMtTom, ane="Anem24bMS", type="boxplot", by="hour")
plotWD (data=wdMtTom, ane="Anem24bMS", type="boxplot", by="month")

## Profile
plotWD (data=wdMtTom, ane=c("Anem24aMS","Anem24bMS", "Anem37aMS" ,"Anem37bMS"), var="ave", type="profile", by="hour")
plotWD (data=wdMtTom,  ane=c("Anem24aMS","Anem24bMS", "Anem37aMS" ,"Anem37bMS"),  var="ave", type="profile", by="month")

## Turbulence Analysis
plotWD (data=wd10, ane="ane10", type="turbulence")

## Fit
plotWD (data=wdMtTom, ane="Anem37aMS", type="fit")

}
\seealso{
\code{\link{tableWD}}, \code{\link{plotwindserie}}, \code{\link{plotcalendar}}
}
\author{
Mariano Bonoli Escobar, Diego Edwards, Valeria Gogni, Ruben Bufanio
}
