% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotwindserie.R
\name{plotwindserie}
\alias{plotwindserie}
\title{Conversion of dataframe in class \code{windata}}
\usage{
plotwindserie(wdata, year, month, ane, var = c("Ave", "Min", "Max", "Temp",
  "Pres", "Dir"), axis = c("Ave", "Min", "Max", "Temp", "Pres", "Dir"),
  shiny = F)
}
\arguments{
\item{wdata}{a dataframe to be converted}

\item{year}{the admit formats are:
- 'YYYYMMDD','YYYY-MM-DD','YYYY.MM.DD' or 'DD/MM/YYYY'}

\item{month}{the name of the variable that contains the times of measurements}

\item{ane}{the name of the variable that contains the dates of measurements}

\item{var}{interval of time betwen two registers. Actually, it's only acepted intervals of 1 minut.}

\item{axis}{the admit formats are:
- 'HHMM','HHMMSS','HH:MM','HH:MM:SS','HH.MM' or 'HH.MM.SS'}

\item{shiny}{the names to indicate the differents anemometers in the list.}
}
\value{
Object of class 'windata' (see details).
}
\description{
Shape a dataframe in a class 'windata'
}
\details{
The object windata is a list with the parameters that were mencionated before.
}
\examples{
# simple example using the windspeed data set
data(data)
 
# let's examine windspeed to see the variables' names
head(data)

}
\author{
Valeria Gogni, Mariano Bonoli, Ruben Bufanio, Diego Edwards
}
