% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotwtg.R
\name{plotwtg}
\alias{plotwtg}
\title{Power curves}
\usage{
plotwtg(datawtg, model)
}
\arguments{
\item{datawtg}{object of class \code{wtgData}.}

\item{model}{one value of character strings with the model wind turbine name. (see details)}
}
\description{
Plot the power curve of differentes wind turbines.
}
\details{
The wind turbines are specificated in \code{wtgData}.
}
\examples{
data("wtgData", package = "WindResource")  
  
## Power curve of wind turbine 'E33'.
plotwtg(wtgData, model = "E33")
}
\author{
Mariano Bonoli Escobar, Diego Edwards, Valeria Gogni, Ruben Bufanio
}
