\name{polar.theme}
\alias{polar.theme}
\title{Extracting paramenters}
\usage{
  polar.theme(plot)
}
\arguments{
  \item{plot}{a list of parameters.}
}
\value{
  The parameter value that was extracted.
}
\description{
  Function to extract a parameter value from a list
}
\author{
  Valeria Gogni, Mariano Bonoli, Ruben Bufanio, Diego
  Edwards
}

