% Generated by roxygen2 (4.0.2): do not edit by hand
\name{powerdensity}
\alias{powerdensity}
\title{Conversion of dataframe in class windata}
\usage{
powerdensity(datawd, ane, dist)
}
\arguments{
\item{datawd}{a dataframe to be converted}

\item{ane}{interval of time betwen two registers. Actually, it's only acepted intervals of 1 minut.}

\item{dist}{the name of the variable that contains the dates of measurements}
}
\value{
Object of class 'windata' (see details).
}
\description{
Shape a dataframe in a class 'windata'
}
\details{
The object windata is a list with the parameters that were mencionated before.
}
\examples{
# simple example using the windspeed data set
data(wd)

# let's examine windspeed to see the variables' names
head(wd)
}
\author{
Valeria Gogni, Mariano Bonoli, Ruben Bufanio, Diego Edwards
}

