% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pref0.R
\name{pref0}
\alias{pref0}
\title{Complete zeros prefix}
\usage{
pref0(data, charnum)
}
\arguments{
\item{data}{a number to complete.}

\item{charnum}{length of target number.}
}
\value{
character number and zeros prefix
}
\description{
Complete zeros prefix
}
\author{
Mariano Bonoli Escobar, Diego Edwards, Valeria Gogni, Ruben Bufanio
}
\keyword{internal}
