% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setWD}
\alias{setWD}
\title{Conversion of dataframe in \code{windata} class object}
\usage{
setWD(data, name = NA, date.var, date.format = c("YYYYMMDD", "YYYY-MM-DD",
  "YYYY.MM.DD", "DD/MM/YYYY"), time.var, time.format = c("HHMM", "HHMMSS",
  "HH:MM", "HH:MM:SS", "HH.MM", "HH.MM.SS"), ane.names, ane.height,
  speed.ave.var, speed.min.var = NA, speed.max.var = NA,
  speed.sd.var = NA, speed.unit = "m/s", dir.var, dir.unit = "deg",
  temp.var = NA, temp.unit = "C", pres.var = NA, pres.unit = "bar",
  NA.values = NA)
}
\arguments{
\item{data}{a dataframe to be converted}

\item{name}{name}

\item{date.var}{name of the variable with the dates of measurements}

\item{date.format}{date format in data: \code{YYYYMMDD}, \code{YYYY-MM-DD}, \code{YYYY.MM.DD} or \code{DD/MM/YYYY}}

\item{time.var}{name of the variable with times of measurements}

\item{time.format}{time format in data: \code{HHMM}, \code{HHMMSS}, \code{HH:MM}, \code{HH:MM:SS}, \code{HH.MM} or \code{HH.MM.SS}}

\item{ane.names}{vector with names of the anemometers.}

\item{ane.height}{vector with the heights of the anemometers in meters. Length of this vector must be the same of ane.names.}

\item{speed.ave.var}{vector with the names of the variables that contains the average speeds measures. Length of this vector must be the same of ane.names.}

\item{speed.min.var}{vector with the names of the variables that contains the minimum measures. Length of this vector must be the same of ane.names.}

\item{speed.max.var}{vector with the names of the variables that contains the maximum measures. Length of this vector must be the same of ane.names.}

\item{speed.sd.var}{vector with the names of the variables that contains the standar desviation measures. Length of this vector must be the same of ane.names.}

\item{speed.unit}{wind speed unit. By the time, only 'm/s' can be used.}

\item{dir.var}{vector with the names of the variables that contains the wind direction measures}

\item{dir.unit}{wind directions unit. By the time, only 'deg' can be used.}

\item{temp.var}{name of the variable that contains the temperature measures.}

\item{temp.unit}{temperature unit.}

\item{pres.var}{name of the variable that contains the pressure measures.}

\item{pres.unit}{pressure unit.}

\item{NA.values}{vector with the the values that must be identified as missing values.}
}
\value{
Object of class \code{windata}.
}
\description{
Shape a dataframe in a class \code{windata} object
}
\details{
The object windata is a list with the parameters that were mencionated before.
}
\examples{
# simple example using the windspeed data set
data(wd)

# let's examine windspeed to see the variables' names
head(wd)
}
\author{
Valeria Gogni, Mariano Bonoli, Ruben Bufanio, Diego Edwards
}

