% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableWD.R
\name{tableWD}
\alias{tableWD}
\title{Summary Tables for \code{Windata} objects}
\usage{
tableWD(datawd, ane = NA, var = c("mean"), type = "histogram",
  by = "none", since = NULL, to = NULL, binwidth = 1)
}
\arguments{
\item{datawd}{an object of class \code{windata}.}

\item{ane}{a vector of character strings with anemometers names to plot.}

\item{var}{a vector of character strings the the variables to plot: \code{ave}, \code{min}, \code{max}, \code{freq}.}

\item{type}{type of plot: \code{histogram}, \code{rose}, \code{profiles} and \code{boxplot}. See also Details.}

\item{by}{an optional string stating if the plot is divided in panels by \code{month} or \code{hour}.}

\item{since}{an optional string with initial date to be taken into account to make the plot. The string format is 'YYYY-MM-DD'.}

\item{to}{an optional string indicating final date to be taken into account to make the plot. The string format is 'YYYY-MM-DD'.}

\item{binwidth}{an optional bindwidth interval used in histogram plots.}
}
\description{
This function returns the values plotted with \code{plotWD} function.
}
\examples{
data("wdMtTom", package = "WindResource")
data("wd10", package = "WindResource")

## Getting anemometer names
wdMtTom$ane$ane.names
wd10$ane$ane.names

## Histogram
tableWD(data=wdMtTom, ane="Anem24bMS", type="histogram", by="none")
tableWD(data=wdMtTom, ane="Anem24bMS", type="histogram", by="none", binwidth=.5)
tableWD(data=wdMtTom, ane="Anem24bMS", type="histogram", by="hour")

## Rose
tableWD (data=wdMtTom, ane=c("Anem24bMS", "Anem37aMS"), var="mean", type="rose", by="none")
tableWD (data=wdMtTom, ane=c("Anem37aMS"), var="mean", type="rose", by="month")

## Boxplot
tableWD (data=wdMtTom, ane="Anem24bMS", type="boxplot", by="hour")
tableWD (data=wdMtTom, ane="Anem24bMS", type="boxplot", by="month")

## Profile
tableWD (data=wdMtTom, ane="Anem24aMS", var="ave", type="profile", by="hour")
tableWD (data=wdMtTom,  ane="Anem24aMS",  var="ave", type="profile", by="month")

## Turbulence Analysis
tableWD (data=wd10, ane="ane10", type="turbulence")

## Fit
tableWD (data=wdMtTom, ane="Anem37aMS", type="fit")

}
\seealso{
\code{\link{plotWD}}
}
\author{
Mariano Bonoli Escobar, Diego Edwards, Valeria Gogni, Ruben Bufanio
}
