% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tablewtg}
\alias{tablewtg}
\title{Plots for winddata objects}
\usage{
tablewtg(data, Model)
}
\arguments{
\item{data}{an object of class winddata}

\item{Model}{an optional vector specifying a subset of anenemometers to plot}
}
\value{
Object of class 'windata' (see details)
}
\description{
Shape a dataframe in a object which will be use by the diferents functions in the package......
}
\details{
The object windata is a list with the parameters that were mencionated before.
}
\examples{
# simple example using the windspeed data set
data(wd)

# let's examine windspeed to see the variables' names
head(wd)
}
\author{
Valeria Gogni, Mariano Bonoli, Ruben Bufanio, Diego Edwards
}

