% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{wd}
\alias{wd}
\title{This is data to be included in my package}
\description{
This is data to be included in my package

This is data to be included in my package
}
\author{
My Name \email{blahblah@roxygen.org}

My Name \email{blahblah@roxygen.org}
}
\references{
\url{data_blah.com}

\url{data_blah.com}
}
\keyword{data}

