% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{wd10}
\alias{wd10}
\title{This is data to be included in my package}
\description{
This is data to be included in my package
}
\author{
My Name \email{blahblah@roxygen.org}
}
\references{
\url{data_blah.com}
}
\keyword{data}

