% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdMtTom.R
\docType{data}
\name{wdMtTom}
\alias{wdMtTom}
\title{wdMtTom}
\description{
Sample data with 10 hour interval
}
\references{
\url{http://www.umass.edu/windenergy/publications/resource/Mt_Tom_Holyoke}
}
\author{
Mariano Bonoli Escobar, Diego Edwards, Valeria Gogni, Ruben Bufanio
}
\keyword{data}
