% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdOlavarria.R
\docType{data}
\name{wdOlavarria}
\alias{wdOlavarria}
\title{wdOlavarria}
\description{
Sample data with 1 hour interval
}
\author{
Mariano Bonoli Escobar, Diego Edwards, Valeria Gogni, Ruben Bufanio
}
\keyword{data}
