% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wt}
\alias{wt}
\title{Extracting paramenters}
\usage{
wt(datawd, datawtg, ane, model)

wt(datawd, datawtg, ane, model)
}
\arguments{
\item{center}{a list of parameters.}

\item{radio}{the parameter value that will be extracted.}

\item{npoints}{the parameter value that will be extracted.}

\item{center}{a list of parameters.}

\item{radio}{the parameter value that will be extracted.}

\item{npoints}{the parameter value that will be extracted.}
}
\value{
The parameter value that was extracted.

The parameter value that was extracted.
}
\description{
Function to extract a parameter value from a list

Function to extract a parameter value from a list
}
\author{
Valeria Gogni, Mariano Bonoli, Ruben Bufanio, Diego Edwards

Valeria Gogni, Mariano Bonoli, Ruben Bufanio, Diego Edwards
}

