% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{wtgData}
\alias{wtgData}
\title{wtgData}
\description{
Tables with differents specifications of wind turbines: \code{"E33"} and \code{"E48"}
}
\author{
Mariano Bonoli Escobar, Diego Edwards, Valeria Gogni, Ruben Bufanio
}
\keyword{data}

