% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add.cart.coord}
\alias{add.cart.coord}
\title{Extracting paramenters}
\usage{
add.cart.coord(dataplot)
}
\arguments{
\item{dataplot}{a list of parameters.}
}
\value{
The parameter value that was extracted.
}
\description{
Function to extract a parameter value from a list
}
\author{
Valeria Gogni, Mariano Bonoli, Ruben Bufanio, Diego Edwards
#'
}

