% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.cart.coord.R
\name{add.cart.coord}
\alias{add.cart.coord}
\title{Add cartesian coordinates to a dataset}
\usage{
add.cart.coord(dataplot)
}
\arguments{
\item{Dataframe}{with value.start, value.end, ang.start and ang.end fields}
}
\value{
The same data.frame entered as parameters, with the following additional fields: x, y, xend and yend
}
\description{
Add cartesian coordinates to a dataset
}
\author{
Mariano Bonoli Escobar, Diego Edwards, Valeria Gogni, Ruben Bufanio
}
\keyword{internal}
