% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check.ane}
\alias{check.ane}
\title{Anemometer Checks}
\usage{
check.ane(datawd, ane)
}
\arguments{
\item{datawd}{an object of class \code{windata}.}

\item{ane}{a vector of character strings with anemometers names.}
}
\value{
a data.frame with the results of annual energy production for different wind turbines models and anemometers.
}
\description{
Calculating the annual energy production of differents wind turbines to differents anemometers.
}
\details{
It's important to see what is the best fit of distribution before to calcuate the annual energy production.
It can be seen with the function \code{plotWD(data, ane, type="fit")}
}
\author{
Valeria Gogni, Mariano Bonoli, Ruben Bufanio, Diego Edwards
}
\seealso{
plotWD
}
\keyword{internal}

