% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circleFun.R
\name{circleFun}
\alias{circleFun}
\title{circle coordinates}
\usage{
circleFun(center = c(0, 0), radio = 1, npoints = 100)
}
\arguments{
\item{center}{vector with the coordinates of the circle center.}

\item{radio}{circle radius.}

\item{npoints}{number of points.}
}
\value{
Data.frame with the circle coordinates.
}
\description{
Generate the
}
\author{
Mariano Bonoli Escobar, Diego Edwards, Valeria Gogni, Ruben Bufanio
}
\keyword{internal}
