% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy.R
\name{energy}
\alias{energy}
\title{Annual energy production calculator}
\usage{
energy(datawd, datawtg, ane, model, dist = NA)
}
\arguments{
\item{datawd}{an object of class \code{windata}.}

\item{datawtg}{object of class \code{wtgData}.}

\item{ane}{a vector of character strings with anemometers names.}

\item{model}{a vector of character strings with the models names of wind turbines.}

\item{dist}{a character string 'name' naming a distribution to calculate the annual energy production. 
The possible distrbutions are: \code{"np"} (no parametric distribution), \code{"weibul"}, \code{"gama"} and \code{"lognormal"}. (see details)}
}
\value{
a data.frame with the results of annual energy production for different wind turbines models and anemometers.
}
\description{
Calculating the annual energy production of differents wind turbines to differents anemometers.
}
\details{
It's important to see what is the best fit of distribution before to calcuate the annual energy production. 
It can be seen with the function \code{plotWD(data, ane, type="fit")}
}
\examples{
data("wd10", package = "WindResource")
data("wtgData", package = "WindResource")

## Getting anemometer names
wd10$ane$ane.names
  
# Calculate 
energy(datawd = wd10, datawtg = wtgData, ane = c("ane10", "ane18"), model = c("E33", "E48"), dist = "weibull")
}
\seealso{
plotWD
}
\author{
Valeria Gogni, Mariano Bonoli, Ruben Bufanio, Diego Edwards
}
