% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fitWD}
\alias{fitWD}
\title{Fit data to selected distribution.}
\usage{
fitWD(data, ane)
}
\arguments{
\item{data}{object of class wd.}

\item{ane}{a vector with the anemometers'name to study. By default 'ane' takes the first anemometer.}
}
\value{
A list with the following components:
-aic: Akaike Index Coefficient.

-loglik: maximun likelihood.

-parameters: a list with the following components:
 name: parameter name
 type: parameter type
 value: estimated parameter
 sd: standard desvest of estimated parameter
}
\description{
Fit of univariate distribution by maximum likelihood.
}
\details{
The 'name' of distribution must be one of 'Weibull', 'Gamma', 'Lognormal'.
}
\examples{
# simple example using the wd class
# list_para <- fitWd(Wd, ane=1, dist='weibull')
}
\author{
Valeria Gogni, Mariano Bonoli, Ruben Bufanio, Diego Edwards
}

