% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotWD}
\alias{plotWD}
\title{Plots for Winddata objects}
\usage{
plotWD(datawd, ane = NA, var = NA, type = c("histogram"), by = c("none",
  "month", "hour"), since = NULL, to = NULL, binwidth = 1)
}
\arguments{
\item{datawd}{an object of class winddata.}

\item{ane}{a vector of character strings containing anemometers names to plot.}

\item{var}{a vector of character strings containing the variables to plot.}

\item{type}{type of plot: 'histogram', 'rose',
'profiles' and 'boxplot'. See also Details.}

\item{by}{an optional string stating if the plot is divided in panels by 'month' or 'hour'.}

\item{since}{an optional string indicating initial date to be taken into account to make the plot.
The string format is 'YYYY-MM-DD'.}

\item{to}{an optional string indicating final date to be taken into account to make the plot.
The string format is 'YYYY-MM-DD'.}
}
\value{
Object of class 'windata' (see details)
}
\description{
Shape a dataframe in a object which will be use by the diferents functions in the package......
}
\details{
The object windata is a list with the parameters that were mencionated before.
}
\examples{
# simple example using the windspeed data set
data(wd)

# let's examine windspeed to see the variables' names
str(wd)
}
\author{
Valeria Gogni, Mariano Bonoli, Ruben Bufanio, Diego Edwards
}

