% Generated by roxygen2 (4.0.2): do not edit by hand
\name{polar.theme}
\alias{polar.theme}
\title{Polar Theme}
\usage{
polar.theme(plot, maxi, by = "none")
}
\arguments{
\item{plot}{}

\item{maxi}{}

\item{by}{}
}
\value{
Polar Theme
}
\description{
Polar Theme used in Rose Plots
}
\author{
Mariano Bonoli Escobar, Diego Edwards, Valeria Gogni, Ruben Bufanio
}
\keyword{internal}

