% Generated by roxygen2 (4.0.1): do not edit by hand
\name{polar.theme}
\alias{polar.theme}
\title{Extracting paramenters}
\usage{
<<<<<<< HEAD
polar.theme(plot, maxi, by = "none")
=======
polar.theme(plot, maxi)
>>>>>>> f66071de49e2956f306270bc4590141274344e01
}
\arguments{
\item{plot}{a list of parameters.}
}
\value{
The parameter value that was extracted.
}
\description{
Function to extract a parameter value from a list
}
\author{
Valeria Gogni, Mariano Bonoli, Ruben Bufanio, Diego Edwards
}

