% Generated by roxygen2 (4.0.2): do not edit by hand
\name{power}
\alias{power}
\title{Extracting paramenters}
\usage{
power(datawd, datawtg = NA, ane, model, dist = "np")
}
\arguments{
\item{datawd}{a list of parameters.}

\item{datawtg}{the parameter value that will be extracted.}

\item{ane}{the parameter value that will be extracted.}

\item{model}{the parameter value that will be extracted.}

\item{dist}{the parameter value that will be extracted.}
}
\value{
The parameter value that was extracted.
}
\description{
Function to extract a parameter value from a list
}
\author{
Valeria Gogni, Mariano Bonoli, Ruben Bufanio, Diego Edwards
}

