\name{pref0}
\alias{pref0}
\title{Complete numbers}
\usage{
  pref0(data, charnum)
}
\arguments{
  \item{data}{a list of paramenters.}

  \item{charnum}{number of digits.}
}
\value{
  parameter's value extracted.
}
\description{
  Function to Complete numbers
}
\author{
  Valeria Gogni, Mariano Bonoli, Ruben Bufanio, Diego
  Edwards
}

