% Generated by roxygen2 (4.0.1): do not edit by hand
\name{runGUI}
\alias{runGUI}
\title{Plots for winddata objects}
\usage{
runGUI(data, inRS = T)
}
\arguments{
\item{data}{a list of parameters.}
}
\description{
Shape a dataframe in a object which will be use by the diferents functions in the package......
}
\details{
The object windata is a list with the parameters that were mencionated before.
}

