\name{runGUI}
\alias{runGUI}
\title{Plots for winddata objects}
\usage{
  runGUI(data)
}
\arguments{
  \item{data}{a list of parameters.}
}
\description{
  Shape a dataframe in a object which will be use by the
  diferents functions in the package......
}
\details{
  The object windata is a list with the parameters that
  were mencionated before.
}

