% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{wdMtTom}
\alias{wdMtTom}
\title{wdMtTom}
\description{
Sample data with 10 hour interval
}
\author{
Mariano Bonoli Escobar, Diego Edwards, Valeria Gogni, Ruben Bufanio
}
\references{
\url{http://www.umass.edu/windenergy/publications/resource/Mt_Tom_Holyoke}
}
\keyword{data}

